+++ crates/deputy/src/main.rs
@@ -1,3 +1,5 @@
+use std::process::ExitCode;
+
 mod cli;
 mod server;
 mod tracing;
@@ -5,7 +7,8 @@ mod tracing;
 use self::tracing::setup_tracing;
 
 #[tokio::main(flavor = "current_thread")]
-async fn main() -> anyhow::Result<()> {
-    setup_tracing();
-    cli::Cli::new().run().await
+async fn main() -> anyhow::Result<ExitCode> {
+    let cli = cli::Cli::new();
+    setup_tracing(cli.quiet_tracing());
+    cli.run().await
 }